added auto-generated relations to entities

+++ entity/src/entities/user_registration.rs
@@ -3,13 +3,12 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-/// A UserRegistration represents the current step that the user has taken in the sign-up flow.
-#[derive(Clone, Debug, Eq, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "user_registration")]
 pub struct Model {
     #[sea_orm(primary_key)]
     #[serde(skip_deserializing)]
-    pub id: i32,
+    pub id: i64,
     pub user_id: i64,
     pub token: Vec<u8>,
     pub status: UserRegistrationStatus,
